adiciona exercício para casa - resolução
