+++ Exercícios/Para casa/README.md
@@ -243,12 +243,12 @@ Teste tudo o que foi criado.
 
 Terminou o exercício? Dá uma olhada nessa checklist e confere se tá tudo certinho, combinado?!
 
-- [ ] Fiz o fork do repositório.
-- [ ] Clonei o fork na minha máquina (`git clone url-do-meu-fork`).
-- [ ] Criei minha branch (`git checkout -b nome-sobrenome`)
-- [ ] Criei a pasta com o meu nome dentro da pasta entregas (` mkdir nome-sobrenome`)
-- [ ] Resolvi o exercício dentro da minha pasta. Como no [exemplo](./entregas/exemplo-nome-sobrenome/).
-- [ ] Adicionei as mudanças. (`git add .` para adicionar todos os arquivos, ou `git add nome_do_arquivo` para adicionar um arquivo específico)
-- [ ] Commitei a cada mudança significativa ou na finalização do exercício (`git commit -m "Mensagem do commit"`)
-- [ ] Pushei os commits na minha branch (`git push origin nome-da-branch`)
-- [ ] Criei um Pull Request seguindo as orientações que estao nesse [documento](./instrucoes-pull-request.md).
+- [OK] Fiz o fork do repositório.
+- [OK] Clonei o fork na minha máquina (`git clone url-do-meu-fork`).
+- [OK] Criei minha branch (`git checkout -b nome-sobrenome`)
+- [OK] Criei a pasta com o meu nome dentro da pasta entregas (` mkdir nome-sobrenome`)
+- [OK] Resolvi o exercício dentro da minha pasta. Como no [exemplo](./entregas/exemplo-nome-sobrenome/).
+- [OK] Adicionei as mudanças. (`git add .` para adicionar todos os arquivos, ou `git add nome_do_arquivo` para adicionar um arquivo específico)
+- [OK] Commitei a cada mudança significativa ou na finalização do exercício (`git commit -m "Mensagem do commit"`)
+- [OK] Pushei os commits na minha branch (`git push origin nome-da-branch`)
+- [OK] Criei um Pull Request seguindo as orientações que estao nesse [documento](./instrucoes-pull-request.md).
