+++ Exercícios/Para casa/entregas/leticia-amaral/Bank.js
@@ -1,0 +1,24 @@
+class Bank {
+  bankCode;
+  bankName;
+  #transferTax;
+
+  static createdBanks = []
+
+  constructor(bankCode, bankName, transferTax) {
+    this.bankCode = bankCode
+    this.bankName = bankName
+    this.#transferTax = transferTax
+    this.constructor.createdBanks.push({ bankCode: this.bankCode, qtdClients: 0})
+  }
+
+  get transferTax() {
+    return this.#transferTax
+  }
+
+  set transferTax(tax) {
+    this.#transferTax = tax
+  }
+}
+
+module.exports = Bank;
